fix build errors
