+++ clients/vault/Cargo.toml
@@ -5,6 +5,10 @@ edition = "2018"
 name = "vault"
 version = "0.0.1"
 
+
+[[examples]]
+name = "oracle"
+
 [features]
 integration = []
 parachain-metadata = ["runtime/parachain-metadata"]
@@ -56,7 +60,7 @@ err-derive = "0.3.1"
 
 [dev-dependencies]
 mockall = "0.8.1"
-
+env_logger = "0.9.0"
 # Workspace dependencies
 runtime = {path = "../runtime", features = ["testing-utils"]}
