+++ clients/vault/src/oracle/constants.rs
@@ -15,9 +15,3 @@ pub const MAX_TXS_PER_FILE: Uint64 = 10_000_000;
 
 /// This is for `TxSetMap`.
 pub const MAX_TXSETS_PER_FILE: usize = 20000;
-
-pub const SAMPLE_VAULT_ADDRESSES_FILTER: &[&str] =
-	&["GAP4SFKVFVKENJ7B7VORAYKPB3CJIAJ2LMKDJ22ZFHIAIVYQOR6W3CXF"];
-
-pub const TIER_1_VALIDATOR_IP_TESTNET: &str = "34.235.168.98";
-pub const TIER_1_VALIDATOR_IP_PUBLIC: &str = "135.181.16.110";
